AttendanceSystem/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2020.12.5=h5b45459_0
certifi=2020.12.5=py38haa244fe_1
click=7.1.2=py_0
dlib=19.21.1=py38h082dfc2_0
dnspython=1.16.0=py38_0
face_recognition=1.3.0=pyh9f0ad1d_1
face_recognition_models=0.3.0=pyh9f0ad1d_0
flask=1.1.2=py_0
flask-sqlalchemy=2.4.4=pyh9f0ad1d_0
freeglut=3.2.1=h0e60522_0
freetype=2.10.4=h546665d_1
icu=67.1=h33f27b4_0
intel-openmp=2020.3=h57928b3_311
itsdangerous=1.1.0=py_0
jasper=2.0.14=h77af90b_2
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9d=h8ffe710_0
libblas=3.9.0=7_mkl
libcblas=3.9.0=7_mkl
libclang=10.0.1=default_hf44288c_1
liblapack=3.9.0=7_mkl
liblapacke=3.9.0=7_mkl
libopencv=4.5.0=py38hf7032e7_7
libpng=1.6.37=h1d00b33_2
libprotobuf=3.13.0.1=h200bbdf_0
libtiff=4.2.0=hc10be44_0
libwebp-base=1.1.0=h8ffe710_3
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
mysql-common=8.0.22=h57928b3_1
mysql-connector-python=8.0.22=py38h7ecd581_0
mysql-libs=8.0.22=hb8c0690_1
numpy=1.19.5=py38h0cc643e_1
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.0=py38haa244fe_7
openssl=1.1.1i=h8ffe710_0
pandas=1.1.3=py38ha925a31_0
pillow=8.1.0=py38hf7ce48b_1
pip=20.2.4=py38_0
protobuf=3.13.0.1=py38ha925a31_1
py-opencv=4.5.0=py38hc5df569_7
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
qt=5.12.9=hb2cf2c5_0
scikit-learn=0.23.2=py38h47e9c7a_0
scipy=1.6.0=py38h5f893b4_0
setuptools=50.3.0=py38h9490d1a_1
six=1.15.0=py_0
sqlalchemy=1.3.22=py38h294d835_1
sqlite=3.33.0=h2a8f88b_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h8ffe710_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_1
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.8=h4e2f164_1
